# docs/api/openapi-mvp.yaml
openapi: 3.0.3
info:
  title: Moimday API
  version: 1.3.0
  description: MVP REST API ??SSOT with api-contract-mvp.md (multi-group, 30 members)
servers:
  - url: http://localhost:3000/v1
    description: Local dev
  - url: https://api.Moimday.example/v1
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /groups:
    post:
      tags: [groups]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
      responses:
        '201': { description: Group created }
        '409': { description: USER_GROUP_LIMIT }
  /groups/{groupId}:
    get:
      tags: [groups]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: groupId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Group + members }
    delete:
      tags: [groups]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Group dissolved (admin) }
  /groups/{groupId}/members/{userId}:
    delete:
      tags: [groups]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Member removed (admin) }
  /groups/{groupId}/admin/transfer:
    post:
      tags: [groups]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Admin transferred }
  /groups/{groupId}/leave:
    post:
      tags: [groups]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Left group }
  /groups/{groupId}/home:
    get:
      tags: [home]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: groupId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Home dashboard }
  /groups/{groupId}/events:
    get:
      tags: [events]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: groupId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: filter
          in: query
          schema: { type: string, enum: [all, my_pending] }
      responses:
        '200': { description: Event list }
    post:
      tags: [events]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: groupId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '201': { description: Event created }
  /groups/{groupId}/calendar:
    get:
      tags: [home]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: groupId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
      responses:
        '200': { description: Calendar entries }
  /groups/{groupId}/invites:
    post:
      tags: [groups]
      security: [{ bearerAuth: [] }]
      responses:
        '201': { description: Invite URL }
  /invites/{token}/accept:
    post:
      tags: [groups]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: token
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Joined group; sets active group }
        '409':
          description: GROUP_FULL | USER_GROUP_LIMIT | ALREADY_MEMBER
...